Vampire deals damage on contact with player, companion projectiles now rotate
