+++ Assets/Scripts/CompanionController.cs.meta
